requirements.txt
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
Flask==1.1.1
Flask-Classful==0.14.2
idna==2.8
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
json5==0.8.4
jsonpickle==1.4.1
MarkupSafe==2.0.1
p2pnetwork==0.0.3
pycryptodome==3.10.4
requests==2.22.0
urllib3==1.26.5
Werkzeug==2.0.2
zipp==3.6.0
